requirements.txt
fastapi==0.104.1
flask==2.3.3
uvicorn==0.23.2
pydantic==1.10.13  # Using an older version that's more compatible with Python 3.13
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0  # Required for direct Supabase API calls
rich==13.6.0
uuid==1.30       # For generating unique IDs